# packages/backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "WallStr.chat",
        "summary": "WallStr.chat OpenAPI schema",
        "version": "0.1.0"
    },
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signup",
                "operationId": "1::signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Signup is disabled"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signin",
                "operationId": "1::signin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessToken"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPUnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Token",
                "operationId": "1::refresh_token",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessToken"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/auth/signout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signout",
                "operationId": "1::signout",
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get Current User",
                "operationId": "1::get_current_user",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/auth/me/settings": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Update User Settings",
                "operationId": "1::update_user_settings",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSettings"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccessToken": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "AccessToken"
            },
...
            "HTTPUnauthorizedError": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "HTTPUnauthorizedError"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SignInRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "SignInRequest"
            },
            "SignUpRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "username": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Username"
                    },
                    "fullname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fullname"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password",
                    "username",
                    "fullname"
                ],
                "title": "SignUpRequest"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "fullname": {
                        "type": "string",
                        "title": "Fullname"
                    },
                    "settings": {
                        "$ref": "#/components/schemas/UserSettings"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "username",
                    "fullname",
                    "settings"
                ],
                "title": "User"
            },
            "UserSettings": {
                "properties": {
                    "llm_model": {
                        "anyOf": [
                            {
                                "type": "string",
                                "const": "claude-3.5-sonnet"
                            },
                            {
                                "type": "string",
                                "const": "deepseek"
                            },
                            {
                                "type": "string",
                                "const": "deepseek-r1"
                            },
                            {
                                "type": "string",
                                "const": "gemini-2.0-flash"
                            },
                            {
                                "type": "string",
                                "const": "gemma-3-27b"
                            },
                            {
                                "type": "string",
                                "const": "gpt-4o"
                            },
                            {
                                "type": "string",
                                "const": "gpt-4o-mini"
                            },
                            {
                                "type": "string",
                                "const": "llama3.1-405b"
                            },
                            {
                                "type": "string",
                                "const": "llama3-70b"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Llm Model"
                    },
                    "simple_mode": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Simple Mode"
                    }
                },
                "type": "object",
                "title": "UserSettings"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}